+++ source/ProcessManager.Abstractions/Api/Model/OrchestrationInstance/ActorIdentityDto.cs
@@ -12,8 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using Energinet.DataHub.ProcessManager.Abstractions.Core.ValueObjects;
+
 namespace Energinet.DataHub.ProcessManager.Abstractions.Api.Model.OrchestrationInstance;
 
 public record ActorIdentityDto(
-    Guid ActorId)
-        : IOperatingIdentityDto;
+    ActorNumber ActorNumber,
+    ActorRole ActorRole)
+    : IOperatingIdentityDto;