+++ source/ProcessManager.Abstractions/Core/ValueObjects/DataHubRecordType.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.ProcessManager.Components.Abstractions.ValueObjects;
+namespace Energinet.DataHub.ProcessManager.Abstractions.Core.ValueObjects;
 
 /// <summary>
 /// Extends enumeration types with strongly typed methods using <typeparamref name="TDataHubRecordType"/>.