+++ source/ProcessManager.Components.Abstractions/ValueObjects/MeteringPointType.cs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 using System.Text.Json.Serialization;
+using Energinet.DataHub.ProcessManager.Abstractions.Core.ValueObjects;
 
 namespace Energinet.DataHub.ProcessManager.Components.Abstractions.ValueObjects;