+++ source/ProcessManager.Core.Tests/Integration/Infrastructure/Orchestration/DurableOrchestrationInstanceExecutorTests.cs
@@ -14,6 +14,7 @@
 
 using DurableTask.Core.Exceptions;
 using Energinet.DataHub.Core.DurableFunctionApp.TestCommon.DurableTask;
+using Energinet.DataHub.ProcessManager.Abstractions.Core.ValueObjects;
 using Energinet.DataHub.ProcessManager.Core.Domain.OrchestrationDescription;
 using Energinet.DataHub.ProcessManager.Core.Domain.OrchestrationInstance;
 using Energinet.DataHub.ProcessManager.Core.Infrastructure.Orchestration;
@@ -192,7 +193,7 @@ public class DurableOrchestrationInstanceExecutorTests : IAsyncLifetime
     {
         var operatingIdentity = new UserIdentity(
             new UserId(Guid.NewGuid()),
-            new ActorId(Guid.NewGuid()));
+            new Actor(ActorNumber.Create("1234567890123"), ActorRole.EnergySupplier));
 
         var orchestrationInstance = OrchestrationInstance.CreateFromDescription(
             operatingIdentity,