+++ source/ProcessManager.Orchestrations.Abstractions/Processes/BRS_026_028/CustomQueries/ActorRequestQuery.cs
@@ -14,6 +14,8 @@
 
 using Energinet.DataHub.ProcessManager.Abstractions.Api.Model;
 using Energinet.DataHub.ProcessManager.Abstractions.Api.Model.OrchestrationInstance;
+using Energinet.DataHub.ProcessManager.Abstractions.Core.ValueObjects;
+using Energinet.DataHub.ProcessManager.Components.Abstractions.ValueObjects;
 using Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Processes.BRS_026_028.BRS_026;
 using Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Processes.BRS_026_028.BRS_028;
 
@@ -34,12 +36,14 @@ public sealed record ActorRequestQuery
     /// <param name="operatingIdentity">Identity of the user executing the query.</param>
     /// <param name="activatedAtOrLater">The time (or later) when the orchestration instances was queued or scheduled to run at.</param>
     /// <param name="activatedAtOrEarlier">The time (or earlier) when the orchestration instances was queued or scheduled to run at.</param>
-    /// <param name="createdByActorId">Optional actor id of the actor to filter by. If not provided, actor requests are returned regardless of who created it.</param>
+    /// <param name="createdByActorNumber">Optional actor number of the actor that created. If not provided, the filter won't be applied.</param>
+    /// <param name="createdByActorRole">Optional actor role of the actor to filter by. If not provided, the filter won't be applied.</param>
     public ActorRequestQuery(
         UserIdentityDto operatingIdentity,
         DateTimeOffset activatedAtOrLater,
         DateTimeOffset activatedAtOrEarlier,
-        Guid? createdByActorId)
+        ActorNumber? createdByActorNumber,
+        ActorRole? createdByActorRole)
             : base(operatingIdentity)
     {
         OrchestrationDescriptionNames = [
@@ -47,7 +51,8 @@ public sealed record ActorRequestQuery
             Brs_028.Name];
         ActivatedAtOrLater = activatedAtOrLater;
         ActivatedAtOrEarlier = activatedAtOrEarlier;
-        CreatedByActorId = createdByActorId;
+        CreatedByActorNumber = createdByActorNumber;
+        CreatedByActorRole = createdByActorRole;
     }
 
     /// <inheritdoc/>
@@ -69,7 +74,12 @@ public sealed record ActorRequestQuery
     public DateTimeOffset ActivatedAtOrEarlier { get; }
 
     /// <summary>
-    /// Optional actor id of the actor to filter by. If not provided, actor requests are returned regardless of who created it.
+    /// Optional actor number of the actor to filter by. If not provided, the filter won't be applied.
     /// </summary>
-    public Guid? CreatedByActorId { get; }
+    public ActorNumber? CreatedByActorNumber { get; }
+
+    /// <summary>
+    /// Optional actor role of the actor to filter by. If not provided, the filter won't be applied.
+    /// </summary>
+    public ActorRole? CreatedByActorRole { get; }
 }