+++ source/ProcessManager.Orchestrations.Tests/Fixtures/OrchestrationsAppFixture.cs
@@ -20,6 +20,8 @@ using Energinet.DataHub.Core.FunctionApp.TestCommon.EventHub.ListenerMock;
 using Energinet.DataHub.Core.FunctionApp.TestCommon.ServiceBus.ListenerMock;
 using Energinet.DataHub.Core.FunctionApp.TestCommon.ServiceBus.ResourceProvider;
 using Energinet.DataHub.Core.TestCommon.Diagnostics;
+using Energinet.DataHub.ProcessManager.Abstractions.Api.Model.OrchestrationInstance;
+using Energinet.DataHub.ProcessManager.Abstractions.Core.ValueObjects;
 using Energinet.DataHub.ProcessManager.Shared.Tests.Fixtures;
 using Microsoft.Azure.WebJobs.Extensions.DurableTask;
 using Xunit.Abstractions;
@@ -109,6 +111,15 @@ public class OrchestrationsAppFixture : IAsyncLifetime
     public ServiceBusListenerMock EnqueueBrs028ServiceBusListener { get; }
 
     public ServiceBusListenerMock IntegrationEventServiceBusListener { get; }
+
+    public ActorIdentityDto DefaultActorIdentity => new ActorIdentityDto(
+        ActorNumber.Create("1234567890123"),
+        ActorRole.EnergySupplier);
+
+    public UserIdentityDto DefaultUserIdentity => new UserIdentityDto(
+        Guid.NewGuid(),
+        DefaultActorIdentity.ActorNumber,
+        DefaultActorIdentity.ActorRole);
 
     private ProcessManagerDatabaseManager DatabaseManager { get; }