+++ source/ProcessManager.Tests/Integration/Processes/BRS_X01/InputExample/V1/MonitorOrchestrationUsingApiScenario.cs
@@ -18,6 +18,8 @@ using System.Text.Json;
 using Energinet.DataHub.Core.TestCommon;
 using Energinet.DataHub.ProcessManager.Abstractions.Api.Model;
 using Energinet.DataHub.ProcessManager.Abstractions.Api.Model.OrchestrationInstance;
+using Energinet.DataHub.ProcessManager.Abstractions.Core.ValueObjects;
+using Energinet.DataHub.ProcessManager.Components.Abstractions.ValueObjects;
 using Energinet.DataHub.ProcessManager.Example.Orchestrations.Abstractions.Processes.BRS_X01.InputExample;
 using Energinet.DataHub.ProcessManager.Example.Orchestrations.Abstractions.Processes.BRS_X01.InputExample.V1;
 using Energinet.DataHub.ProcessManager.Example.Orchestrations.Abstractions.Processes.BRS_X01.InputExample.V1.Model;
@@ -35,6 +37,11 @@ namespace Energinet.DataHub.ProcessManager.Tests.Integration.Processes.BRS_X01.I
 [Collection(nameof(ProcessManagerAppCollection))]
 public class MonitorOrchestrationUsingApiScenario : IAsyncLifetime
 {
+    private readonly UserIdentityDto _userIdentity = new UserIdentityDto(
+        UserId: Guid.NewGuid(),
+        ActorNumber: ActorNumber.Create("1234567890123"),
+        ActorRole: ActorRole.EnergySupplier);
+
     public MonitorOrchestrationUsingApiScenario(
         ProcessManagerAppFixture fixture,
         ITestOutputHelper testOutputHelper)
@@ -64,16 +71,12 @@ public class MonitorOrchestrationUsingApiScenario : IAsyncLifetime
     [Fact]
     public async Task ExampleOrchestration_WhenStarted_CanMonitorLifecycle()
     {
-        var userIdentity = new UserIdentityDto(
-            UserId: Guid.NewGuid(),
-            ActorId: Guid.NewGuid());
-
         var orchestration = Brs_X01_InputExample.V1;
         var input = new InputV1(
             ShouldSkipSkippableStep: false);
 
         var command = new StartInputExampleCommandV1(
-             operatingIdentity: userIdentity,
+             operatingIdentity: _userIdentity,
              input);
 
         using var startRequest = new HttpRequestMessage(
@@ -95,7 +98,7 @@ public class MonitorOrchestrationUsingApiScenario : IAsyncLifetime
 
         // Step 2: Query until terminated with succeeded
         var query = new GetOrchestrationInstanceByIdQuery(
-            userIdentity,
+            _userIdentity,
             orchestrationInstanceId);
 
         var isTerminated = await Awaiter.TryWaitUntilConditionAsync(
@@ -133,17 +136,13 @@ public class MonitorOrchestrationUsingApiScenario : IAsyncLifetime
     [Fact]
     public async Task ExampleOrchestration_WhenScheduledToRunNow_CanSearch()
     {
-        var userIdentity = new UserIdentityDto(
-            UserId: Guid.NewGuid(),
-            ActorId: Guid.NewGuid());
-
         var now = DateTimeOffset.UtcNow;
         var orchestration = Brs_X01_InputExample.V1;
         var input = new InputV1(
             ShouldSkipSkippableStep: false);
 
         var command = new ScheduleInputExampleCommandV1(
-             operatingIdentity: userIdentity,
+             operatingIdentity: _userIdentity,
              input,
              runAt: now);
 
@@ -166,7 +165,7 @@ public class MonitorOrchestrationUsingApiScenario : IAsyncLifetime
 
         // Step 2: General search using name
         var queryByName = new SearchOrchestrationInstancesByNameQuery(
-            userIdentity,
+            _userIdentity,
             orchestration.Name,
             version: null,
             lifecycleState: null,
@@ -195,7 +194,7 @@ public class MonitorOrchestrationUsingApiScenario : IAsyncLifetime
 
         // Step 3: Custom search
         var customQuery = new InputExampleQuery(
-            userIdentity,
+            _userIdentity,
             skippedStepTwo: input.ShouldSkipSkippableStep);
 
         using var customQueryRequest = new HttpRequestMessage(