+++ source/ProcessManager.Tests/Integration/Scheduler/SchedulerHandlerTests.cs
@@ -13,6 +13,8 @@
 // limitations under the License.
 
 using Energinet.DataHub.Core.App.Common.Extensions.DependencyInjection;
+using Energinet.DataHub.ProcessManager.Abstractions.Core.ValueObjects;
+using Energinet.DataHub.ProcessManager.Components.Abstractions.ValueObjects;
 using Energinet.DataHub.ProcessManager.Core.Application.Orchestration;
 using Energinet.DataHub.ProcessManager.Core.Application.Registration;
 using Energinet.DataHub.ProcessManager.Core.Domain.OrchestrationDescription;
@@ -52,7 +54,7 @@ public class SchedulerHandlerTests : IClassFixture<SchedulerHandlerFixture>, IAs
         _now = _fixture.ClockMock.Object.GetCurrentInstant();
         _userIdentity = new UserIdentity(
             new UserId(Guid.NewGuid()),
-            new ActorId(Guid.NewGuid()));
+            new Actor(ActorNumber.Create("1234567890123"), ActorRole.EnergySupplier));
 
         _executorMock = new Mock<IOrchestrationInstanceExecutor>();